refactor(deep_causality): Decluttered tests for EffectEthos
