+++ deep_causality/tests/types/telos_types/effect_ethos/mod.rs
@@ -2,4 +2,11 @@
  * SPDX-License-Identifier: MIT
  * Copyright (c) "2025" . The DeepCausality Authors and Contributors. All Rights Reserved.
  */
-mod effect_ethos_tests;
+mod effect_ethos_evaluate_action_tests;
+mod effect_ethos_freeze_tests;
+mod effect_ethos_linking_tests;
+mod effect_ethos_norm_tests;
+mod effect_ethos_constructor_tests;
+mod effect_ethos_verdict_tests;
+mod effect_ethos_verify_graph_tests;
+mod utils_tests;
